+++ pages/fotos/index.tsx
